backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Flask-Login==0.6.3
authlib==1.2.1
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gunicorn==21.2.0
python-dateutil==2.8.2 